# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TUP Open Data API",
    "version": "0.1.0"
  },
  "components": {
    "schemas": {
...
      "Provenance": {
        "type": "object",
        "properties": {
          "source_url": {
            "type": "string",
            "description": "Where this record came from. A citable URL, or seed:// for curated rows."
          },
          "first_seen_at": {
            "type": "string",
            "format": "date-time",
            "description": "When this record was first ingested."
          },
          "last_verified_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the source was last confirmed to still say this. NOT when TUP updated it."
          },
          "staleness_days": {
            "type": "integer",
            "description": "Whole days since last_verified_at."
          },
          "confidence": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "description": "How much this record should be trusted. Ascending: low < medium < high. Derived from source freshness and entity type, not from parser certainty.",
            "example": "high"
          },
          "method": {
            "type": "string",
            "description": "crawl | manual | partner_feed | seed",
            "example": "seed"
          }
        },
        "required": [
          "source_url",
          "first_seen_at",
          "last_verified_at",
          "staleness_days",
          "confidence",
          "method"
        ]
      },
      "CollectionMeta": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "description": "Number of items in this page."
          },
          "has_more": {
            "type": "boolean",
            "description": "Authoritative signal for \"keep paginating\". Prefer over links.next."
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "description": "When this response was assembled. Independent of data age — read provenance.last_verified_at for that. Deliberately excluded from the ETag."
          },
          "freshness": {
            "$ref": "#/components/schemas/Freshness"
          }
        },
        "required": [
          "count",
          "has_more",
          "generated_at",
          "freshness"
        ]
      },
      "Freshness": {
        "type": "object",
        "properties": {
          "oldest_verified_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "The least recently verified record in this page."
          },
          "max_staleness_days": {
            "type": "integer"
          },
          "min_confidence": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "low",
              "medium",
              "high",
              null
            ],
            "description": "How much this record should be trusted. Ascending: low < medium < high. Derived from source freshness and entity type, not from parser certainty.",
            "example": "high"
          },
          "counts_by_confidence": {
            "type": "object",
            "properties": {
              "high": {
                "type": "integer"
              },
              "medium": {
                "type": "integer"
              },
              "low": {
                "type": "integer"
              }
            },
            "required": [
              "high",
              "medium",
              "low"
            ]
          }
        },
        "required": [
          "oldest_verified_at",
          "max_staleness_days",
          "min_confidence",
          "counts_by_confidence"
        ]
      },
      "CollectionLinks": {
        "type": "object",
        "properties": {
          "self": {
            "type": "string"
          },
          "next": {
            "type": [
              "string",
              "null"
            ],
            "description": "Opaque. Do not decode or construct cursors — the format may change."
          }
        },
        "required": [
          "self",
          "next"
        ]
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "URI identifying the error class. Resolves to docs."
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string",
            "description": "Names the offending value, not just the class."
          },
          "instance": {
            "type": "string"
          },
          "did_you_mean": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Trigram-similar identifiers, on 404 for a slug-shaped path."
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ]
      },
...
      "Program": {
        "type": "object",
        "properties": {
          "ref": {
            "type": "string",
            "example": "bsce"
          },
          "slug": {
            "type": "string",
            "example": "bsce"
          },
          "code": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string",
            "example": "Bachelor of Science in Civil Engineering"
          },
          "aliases": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Source name variants the matching chain accepts. Never `null`."
          },
          "level": {
            "type": "string",
            "enum": [
              "certificate",
              "diploma",
              "associate",
              "baccalaureate",
              "masters",
              "doctorate",
              "post_baccalaureate"
            ],
            "example": "baccalaureate"
          },
          "discipline": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "typical_years": {
            "type": [
              "number",
              "null"
            ]
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          }
        },
        "required": [
          "ref",
          "slug",
          "code",
          "name",
          "aliases",
          "level",
          "discipline",
          "description",
          "typical_years",
          "provenance"
        ]
      },
...
      "ProgramOffering": {
        "type": "object",
        "properties": {
          "ref": {
            "type": "string",
            "example": "manila/bsce"
          },
          "campus": {
            "type": "string",
            "enum": [
              "manila",
              "cavite",
              "visayas",
              "taguig"
            ],
            "description": "Campus identifier.",
            "example": "manila"
          },
          "slug": {
            "type": "string",
            "example": "bsce"
          },
          "program": {
            "type": [
              "string",
              "null"
            ],
            "description": "Canonical program slug, or null when the source name has not been resolved into the registry. Never auto-created from a fuzzy match (ADR-003).",
            "example": "bsce"
          },
          "source_name": {
            "type": "string",
            "description": "Verbatim, as the campus publishes it, before canonicalisation.",
            "example": "Bachelor of Science in Civil Engineering"
          },
          "local_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "unit": {
            "$ref": "#/components/schemas/OfferingUnit"
          },
          "majors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Never `null`."
          },
          "years": {
            "type": [
              "number",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "suspended",
              "phased_out",
              "unknown",
              "removed"
            ],
            "example": "active"
          },
          "curriculum_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "accreditation": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {}
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          }
        },
        "required": [
          "ref",
          "campus",
          "slug",
          "program",
          "source_name",
          "local_name",
          "unit",
          "majors",
          "years",
          "status",
          "curriculum_url",
          "accreditation",
          "provenance"
        ]
      },
      "OfferingUnit": {
        "type": [
          "object",
          "null"
        ],
        "properties": {
          "ref": {
            "type": "string",
            "example": "manila/coe"
          },
          "slug": {
            "type": "string",
            "example": "coe"
          },
          "name": {
            "type": "string",
            "example": "College of Engineering"
          },
          "type": {
            "type": "string",
            "enum": [
              "college",
              "department",
              "institute",
              "center",
              "program_group"
            ],
            "description": "Organisational vocabulary of this unit. Campuses differ: Manila and Visayas use colleges, Cavite uses departments. Always read this field; never assume \"college\".",
            "example": "college"
          }
        },
        "required": [
          "ref",
          "slug",
          "name",
          "type"
        ]
      }
    },
...
  },
  "paths": {
...
    "/v1/programs": {
      "get": {
        "tags": [
          "programs"
        ],
        "summary": "List canonical degree programs",
        "description": "The campus-agnostic degree registry (ADR-003). `campus`, `unit` and `status` are filters over the *offerings* of each program — a program is not owned by a campus, which is why there is no `/v1/campuses/{campus}/programs` (docs/13 §2.3).",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "manila",
                "cavite",
                "visayas",
                "taguig"
              ],
              "description": "Restrict to one campus. Validated against the four known slugs.",
              "example": "manila"
            },
            "required": false,
            "description": "Restrict to one campus. Validated against the four known slugs.",
            "name": "campus",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "certificate",
                "diploma",
                "associate",
                "baccalaureate",
                "masters",
                "doctorate",
                "post_baccalaureate"
              ]
            },
            "required": false,
            "name": "level",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Unit ref, e.g. `manila/coe`. Matches programs offered by that unit.",
              "example": "manila/coe"
            },
            "required": false,
            "description": "Unit ref, e.g. `manila/coe`. Matches programs offered by that unit.",
            "name": "unit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "example": "engineering"
            },
            "required": false,
            "name": "discipline",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "suspended",
                "phased_out",
                "unknown",
                "removed"
              ],
              "description": "Status of the offering used to match `campus`/`unit`."
            },
            "required": false,
            "description": "Status of the offering used to match `campus`/`unit`.",
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200,
              "description": "Case-insensitive substring of the program name.",
              "example": "civil"
            },
            "required": false,
            "description": "Case-insensitive substring of the program name.",
            "name": "q",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "low",
                "medium",
                "high"
              ],
              "description": "Inclusive lower bound: `confidence >= value`. Default is no filter — the API returns everything and labels it, because hiding low-confidence rows leaves a consumer unable to tell \"none exist\" from \"none we vouch for\".",
              "example": "medium"
            },
            "required": false,
            "description": "Inclusive lower bound: `confidence >= value`. Default is no filter — the API returns everything and labels it, because hiding low-confidence rows leaves a consumer unable to tell \"none exist\" from \"none we vouch for\".",
            "name": "min_confidence",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "description": "Page size. Default 20, maximum 100; above that is a 400.",
              "example": 20
            },
            "required": false,
            "description": "Page size. Default 20, maximum 100; above that is a 400.",
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Opaque continuation token from links.next. Do not decode or construct one — the format may change without a version bump."
            },
            "required": false,
            "description": "Opaque continuation token from links.next. Do not decode or construct one — the format may change without a version bump.",
            "name": "cursor",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "A page of canonical programs.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Program"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/CollectionMeta"
                    },
                    "links": {
                      "$ref": "#/components/schemas/CollectionLinks"
                    }
                  },
                  "required": [
                    "data",
                    "meta",
                    "links"
                  ]
                }
              }
            }
          },
          "304": {
            "description": "Not modified."
          },
          "400": {
            "description": "Unknown parameter, limit above 100, or a mismatched cursor.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/offerings": {
      "get": {
        "tags": [
          "programs"
        ],
        "summary": "List degrees as each campus actually publishes them",
        "description": "One row per degree per campus, carrying the verbatim `source_name`. This is the only view that includes offerings whose source name has not yet been resolved into the canonical registry (`program: null`) — they are real degrees a campus teaches, and hiding them until a human has classified them would misrepresent the catalogue. Filter them with `program=unmatched`.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "manila",
                "cavite",
                "visayas",
                "taguig"
              ],
              "description": "Restrict to one campus. Validated against the four known slugs.",
              "example": "manila"
            },
            "required": false,
            "description": "Restrict to one campus. Validated against the four known slugs.",
            "name": "campus",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Unit ref, e.g. `manila/coe`."
            },
            "required": false,
            "description": "Unit ref, e.g. `manila/coe`.",
            "name": "unit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Canonical program slug. Use `unmatched` for offerings not yet resolved.",
              "example": "bsce"
            },
            "required": false,
            "description": "Canonical program slug. Use `unmatched` for offerings not yet resolved.",
            "name": "program",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "suspended",
                "phased_out",
                "unknown",
                "removed"
              ]
            },
            "required": false,
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200,
              "description": "Case-insensitive substring of the verbatim source name."
            },
            "required": false,
            "description": "Case-insensitive substring of the verbatim source name.",
            "name": "q",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "low",
                "medium",
                "high"
              ],
              "description": "Inclusive lower bound: `confidence >= value`. Default is no filter — the API returns everything and labels it, because hiding low-confidence rows leaves a consumer unable to tell \"none exist\" from \"none we vouch for\".",
              "example": "medium"
            },
            "required": false,
            "description": "Inclusive lower bound: `confidence >= value`. Default is no filter — the API returns everything and labels it, because hiding low-confidence rows leaves a consumer unable to tell \"none exist\" from \"none we vouch for\".",
            "name": "min_confidence",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "description": "Page size. Default 20, maximum 100; above that is a 400.",
              "example": 20
            },
            "required": false,
            "description": "Page size. Default 20, maximum 100; above that is a 400.",
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Opaque continuation token from links.next. Do not decode or construct one — the format may change without a version bump."
            },
            "required": false,
            "description": "Opaque continuation token from links.next. Do not decode or construct one — the format may change without a version bump.",
            "name": "cursor",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "A page of offerings.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProgramOffering"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/CollectionMeta"
                    },
                    "links": {
                      "$ref": "#/components/schemas/CollectionLinks"
                    }
                  },
                  "required": [
                    "data",
                    "meta",
                    "links"
                  ]
                }
              }
            }
          },
          "304": {
            "description": "Not modified."
          },
          "400": {
            "description": "Unknown parameter, limit above 100, or a mismatched cursor.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}